Nation cleanup
